requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.3
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
stripe==13.0.0
boto3==1.40.43
django-storages==1.14.6
django-ratelimit==4.1.0
django-csp==3.7
django-security==0.20.0
drf-spectacular==0.26.5
Pillow==10.1.0
